Split visitor logic from activation handler logic
